Add the improved and rewritten version of Gruvlinan
This version is much cleaned upp from the C++ counterpart and it has a lot of improvements. It doesn't print unnecessary information, a lot of the bigger calculations are moved out to external functions along with materials being chosen in a sane way. The materials use object oriented programming in order to simplify material values.