Cleanups and fixes for improved tic-tac-toe
We now dropped the buffered channel because we don't need it, tell the buttons if we are in a game or not and we clean up the player and press booleans on each button. This cleanup is done on a separate goroutine just after winning the match in order to make it plenty fast. Some unnecessary case checking in two places where also dropped.
These changes should improve the speed, but not a noticeable amount.